pyproject.toml
[project]
name = "triage-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "mcp>=1.13.1",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
    "google-api-python-client>=2.129.0",
    "google-auth>=2.30.0",
    "google-auth-httplib2>=0.2.0",
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
]


[project.scripts]
client = "client:run_chat"
server = "server:main"
api = "api:main" 